chore: Test 관련 Configuration 통일 및 Github API 관련 클래스 프로필 설정 (#217)
