+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/PickGitApplicationTests.java
@@ -1,12 +1,12 @@
 package com.woowacourse.pickgit;
 
-import com.woowacourse.pickgit.unit.post.PostTestConfiguration;
+import com.woowacourse.pickgit.config.TestApiConfiguration;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 
-@Import(PostTestConfiguration.class)
+@Import(TestApiConfiguration.class)
 @SpringBootTest
 @ActiveProfiles("test")
 class PickGitApplicationTests {
