+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/config/TestApiConfiguration.java
@@ -1,0 +1,29 @@
+package com.woowacourse.pickgit.config;
+
+import com.woowacourse.pickgit.common.mockapi.MockPickGitStorage;
+import com.woowacourse.pickgit.common.mockapi.MockRepositoryApiRequester;
+import com.woowacourse.pickgit.common.mockapi.MockTagApiRequester;
+import com.woowacourse.pickgit.post.domain.PickGitStorage;
+import com.woowacourse.pickgit.post.infrastructure.PlatformRepositoryApiRequester;
+import com.woowacourse.pickgit.tag.infrastructure.PlatformTagApiRequester;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+
+@TestConfiguration
+public class TestApiConfiguration {
+
+    @Bean
+    public PlatformRepositoryApiRequester platformRepositoryApiRequester() {
+        return new MockRepositoryApiRequester();
+    }
+
+    @Bean
+    public PlatformTagApiRequester platformTagApiRequester() {
+        return new MockTagApiRequester();
+    }
+
+    @Bean
+    public PickGitStorage pickGitStorage() {
+        return new MockPickGitStorage();
+    }
+}
