+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/config/TestJpaConfiguration.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pickgit;
+package com.woowacourse.pickgit.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
@@ -6,5 +6,4 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 @TestConfiguration
 @EnableJpaAuditing
 public class TestJpaConfiguration {
-
 }
